pyproject.toml
[project]
name = "ts-mod1-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "sentence-transformers>=5.1.0",
]
